Move models to `v1alpha` package
As we're going to have lots of packages in the future, we should aim to
use a package per version.

Closes #57.
