+++ examples/java/spring-boot/src/main/java/uk/gov/api/springboot/controllers/MetadataController.java
@@ -4,8 +4,8 @@ import java.util.List;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.api.models.metadata.ApiMetadata;
-import uk.gov.api.models.metadata.BulkMetadataResponse;
+import uk.gov.api.models.metadata.v1alpha.ApiMetadata;
+import uk.gov.api.models.metadata.v1alpha.BulkMetadataResponse;
 import uk.gov.api.springboot.services.MetadataService;
 
 @RestController
