+++ examples/java/spring-boot/src/test/java/uk/gov/api/springboot/controllers/MetadataControllerIntegrationTest.java
@@ -20,8 +20,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
-import uk.gov.api.models.metadata.ApiMetadata;
-import uk.gov.api.models.metadata.Data;
+import uk.gov.api.models.metadata.v1alpha.ApiMetadata;
+import uk.gov.api.models.metadata.v1alpha.Data;
 import uk.gov.api.springboot.services.MetadataService;
 
 @AutoConfigureMockMvc
