+++ examples/java/spring-boot/src/test/java/uk/gov/api/springboot/controllers/MetadataControllerTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.api.models.metadata.ApiMetadata;
-import uk.gov.api.models.metadata.BulkMetadataResponse;
+import uk.gov.api.models.metadata.v1alpha.ApiMetadata;
+import uk.gov.api.models.metadata.v1alpha.BulkMetadataResponse;
 import uk.gov.api.springboot.services.MetadataService;
 
 @ExtendWith(MockitoExtension.class)
